more clean up work, separate files for follower and leader
