+++ raft/consensus_module.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"log"
 	"math/rand"
-	"net/rpc"
 	"sync"
 	"time"
 )
@@ -48,100 +47,6 @@ type AppendEntriesReply struct {
 func (cm *ConsensusModule) log(format string, args ...interface{}) {
 	format = fmt.Sprintf("[ %v ]\t %v \t", cm.server.id, cm.server.state) + format
 	log.Printf(format, args...)
-}
-
-func (cm *ConsensusModule) ChangeState(nextState string) {
-	cm.mu.Lock()
-
-	if cm.server.state == nextState { // verify
-		// if state transition already done (unique scenario, then ignore the rest of the code, unlock the mutex and leave)
-		// this scenario arise when, candidate already moved back to follower state upon receive negative ack from followers
-		// but state change was triggered again due to negative acknowledgement of a competing candidate (with higher term obviously, hence the negative ack)
-		cm.mu.Unlock()
-		return
-	}
-	cm.server.state = nextState
-
-	if nextState == LEADER {
-		cm.doWhatALeaderDoes()
-	} else if nextState == FOLLOWER {
-		cm.log("Became a follower")
-		// run election timer
-	}
-}
-
-func (cm *ConsensusModule) becomeLeader() {
-	// cm.mu.Lock()
-	cm.doWhatALeaderDoes()
-}
-
-func (cm *ConsensusModule) becomeFollower() {
-	cm.mu.Lock()
-	cm.server.state = FOLLOWER
-	cm.mu.Unlock()
-}
-
-func (cm *ConsensusModule) sendHeartbeatToPeer(peerId string, peerClient *rpc.Client, wg *sync.WaitGroup, args AppendEntriesArgs, reply *AppendEntriesReply) {
-	defer cm.wg.Done()
-
-	cm.log("Sending heartbeat to %v", peerId)
-	args.To = peerId
-
-	if err := peerClient.Call("ConsensusModule.AppendEntries", args, reply); err == nil {
-		if reply.Success {
-			cm.log("Still the leader, yayyy!")
-		} else {
-			cm.currentTerm = reply.Term
-			cm.becomeFollower()
-			cm.done <- true
-		}
-		// cm.log("Heartbeat response from peer %v is %v ", peerId, reply.Success)
-	}
-}
-
-func (cm *ConsensusModule) sendHeartbeats(periodic bool) {
-	defer cm.wg.Done()
-
-	args := AppendEntriesArgs{
-		From: cm.server.id,
-		Term: cm.currentTerm,
-	}
-	reply := AppendEntriesReply{}
-	if !periodic {
-		for peerId, peerClient := range cm.server.peerClients {
-			cm.wg.Add(1)
-			go cm.sendHeartbeatToPeer(peerId, peerClient, &cm.wg, args, &reply)
-		}
-		return
-	}
-	cm.ticker = time.NewTicker(2 * time.Second)
-	for {
-		select {
-		case <-cm.done:
-			cm.ticker.Stop()
-			return
-		case <-cm.ticker.C:
-			for peerId, peerClient := range cm.server.peerClients {
-				cm.wg.Add(1)
-				go cm.sendHeartbeatToPeer(peerId, peerClient, &cm.wg, args, &reply)
-			}
-		}
-	}
-}
-
-// expect cm.mu to be locked
-func (cm *ConsensusModule) doWhatALeaderDoes() {
-	cm.log("Became a leader, sending heartbeats")
-
-	// this is the first heartbeat after becoming the leader, not sure if I should wait for it to finish
-	cm.wg.Add(1) // I am not sure if I need this
-	go cm.sendHeartbeats(false)
-	cm.wg.Wait() // I am not sure if I need this
-
-	cm.wg.Add(1)
-	go cm.sendHeartbeats(true)
-	cm.wg.Wait()
-	cm.mu.Unlock()
 }
 
 // procedures
